Use a bigger tolerance to determine if it is a hairline
fast_len() used to determine whether a stroke is a hairline is an
approx estimate.  It avoids sqrt operation, but gives false negative for
certain case.

We make relaxed computation - use 1.1 instead of 1.0 for tolerance.  We
still will have false negative.  For example (0.9, 0.424) will be 1.112.
But chances of false negative is smaller.

The result is that some of previously recognized as non-hariline is now.
hairline, and uses different renderer.

For example, for non-msaa target, hairline is rendered via
GrAAhairLinePathRenderer, while non-hairline is renderered via
GrAAConvexPathRenderer/GrSoftwarePathRenderer.

For msaa target, hairline is renderered via GrHairLinePathRenderer, while
non hairline is renderred via GrShapePathRenderer.