Clip optimization for GL msaa render target
When calling clipPath/clipRegion/clipRect/clipRRect, skia scans the entire
clip path to generate an accurate clip bounds. This scans one line at a time.

It is useful for image render target or non-msaa GL render target where it
needs to generate clip mask in a bitmap.  However, for GL msaa render target,
the draw control flows never hit the software path.  So such scanline is a
waste.  The optimization only computes the clip bounding rectangles, and
therefore is much faster.

We add a new API on SkBaseDevice::renderTargetHasMultisampling().  This is
because in SkDeferredCanvas, we cannot get GrRenderTarget without first
flush the pending drawing.  This causes problem in query for isFreshFrame().
If we flush pending commands, query for isFreshFrame() returns existing
fFreshFrame value while witout flush pending, skipPendingCommands() will reset
fFreshFrame = true;

There is additional benefit of this API, in case of clipping, paraellel canvas
will not fall back to non-parallel mode